Refactor Supabase integration
