+++ src/app/[collectiveSlug]/[postId]/page.tsx
@@ -1,5 +1,4 @@
-import { cookies } from "next/headers";
-import { createServerClient } from "@supabase/ssr";
+import { createServerSupabaseClient } from "@/lib/supabase/server";
 import type { Database } from "@/lib/database.types";
 import { notFound } from "next/navigation";
 import { Button } from "@/components/ui/button";
@@ -45,12 +44,7 @@ export default async function PostPage({
   params: Promise<{ collectiveSlug: string; postId: string }>;
 }) {
   const { collectiveSlug, postId } = await params;
-  const cookieStore = await cookies();
-  const supabase = createServerClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    { cookies: cookieStore }
-  );
+  const supabase = await createServerSupabaseClient();
 
   const {
     data: { user },
