+++ src/app/actions/memberActions.ts
@@ -1,8 +1,7 @@
 "use server";
 
 import { z } from "zod";
-import { cookies } from "next/headers";
-import { createServerClient } from "@supabase/ssr";
+import { createServerSupabaseClient } from "@/lib/supabase/server";
 import type { Database, Enums } from "@/lib/database.types";
 import {
   InviteMemberServerSchema,
@@ -23,18 +22,7 @@ interface ActionResult<T = null> {
 export async function inviteMemberToCollective(
   formData: InviteMemberServerValues
 ): Promise<ActionResult> {
-  const cookieStore = await cookies();
-  const supabase = createServerClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get: (name) => cookieStore.get(name)?.value,
-        set: () => {},
-        remove: () => {},
-      },
-    }
-  );
+  const supabase = await createServerSupabaseClient();
 
   const {
     data: { user: currentUser },
