+++ src/app/api/subscribe/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { cookies } from "next/headers";
-import { createServerClient, type CookieOptions } from "@supabase/ssr";
+import { createRouteHandlerClient } from "@supabase/auth-helpers-nextjs";
 import { getStripe } from "@/lib/stripe";
 import Stripe from "stripe"; // Explicitly import Stripe namespace
 import { supabaseAdmin } from "@/lib/supabaseAdmin"; // Updated import path
@@ -23,34 +23,7 @@ interface StripeErrorLike {
 }
 
 export async function POST(request: Request) {
-  const cookieStore = await cookies();
-
-  // Using createServerClient instead of createRouteHandlerClient
-  const supabase = createServerClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        set(_name: string, _value: string, _options: CookieOptions) {
-          // For Route Handlers, NextRequest's cookies are read-only.
-          // To set cookies, you must return them in the NextResponse.
-          // This simplified client setup is for auth.getSession() primarily.
-          // If setting cookies directly via Supabase client in a POST/PUT is needed,
-          // this part requires more complex handling with NextResponse.
-          // console.warn('Supabase client attempted to set cookie in Route Handler via createServerClient.');
-        },
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        remove(_name: string, _options: CookieOptions) {
-          // Similar to set, direct removal isn't standard here without NextResponse manipulation.
-          // console.warn('Supabase client attempted to remove cookie in Route Handler via createServerClient.');
-        },
-      },
-    }
-  );
+  const supabase = createRouteHandlerClient<Database>({ cookies });
 
   const stripe = getStripe();
   if (!stripe) {
