+++ src/components/app/dashboard/nav/DashboardNav.tsx
@@ -7,7 +7,8 @@ import type { CollectiveSummary } from "../template/dashboard-shell";
 import React, { useState, useEffect } from "react";
 import { CollectiveSelectorDropdown } from "../molecules/CollectiveSelectorDropdown";
 import { UserMenu } from "../molecules/UserMenu";
-import { createBrowserClient } from "@supabase/ssr";
+import { createBrowserClient } from "@supabase/auth-helpers-nextjs";
+import type { Database } from "@/lib/database.types";
 import { SidebarNav } from "../molecules/SidebarNav";
 import * as Sheet from "@radix-ui/react-dialog";
 
@@ -42,10 +43,7 @@ export function DashboardNav({
   const [sheetOpen, setSheetOpen] = useState<boolean>(false);
 
   useEffect(() => {
-    const supabase = createBrowserClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL!,
-      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-    );
+    const supabase = createBrowserClient<Database>();
     supabase.auth.getUser().then(async ({ data }: { data: any }) => {
       if (data.user) {
         const { id, email } = data.user;
