--- src/lib/auth.ts
@@ -1,1 +1,0 @@
-
