+++ src/lib/supabase/browser.ts
@@ -1,12 +1,9 @@
-import { createBrowserClient } from "@supabase/ssr";
+import { createBrowserClient } from "@supabase/auth-helpers-nextjs";
 import type { Database } from "../database.types";
 import { SupabaseClient } from "@supabase/supabase-js";
 
 export const createSupabaseBrowserClient = (): SupabaseClient<Database> =>
-  createBrowserClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  );
+  createBrowserClient<Database>();
 
 const supabase = createSupabaseBrowserClient();
 export default supabase;
