--- src/types/@supabase-ssr.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-declare module "@supabase/ssr" {
-  export const createBrowserClient: any;
-  export const createServerClient: any;
-}
